requirements.txt
antlr4-python3-runtime==4.8
asgiref==3.4.1
bitarray==2.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
Cython==0.29.32
defusedxml==0.7.1
Django==3.2.15
django-cors-headers==3.10.1
django-environ==0.9.0
django-extensions==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
fairseq==0.12.2
gunicorn==20.1.0
hydra-core==1.0.7
idna==3.3
importlib-resources==5.4.0
lxml==4.9.1
numpy==1.19.5
omegaconf==2.0.6
portalocker==2.5.1
pycparser==2.21
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.27.1
sacrebleu==2.2.0
sqlparse==0.4.2
tabulate==0.8.10
torch==1.10.1
torchaudio==0.10.1
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.12
zipp==3.6.0
